+++ src/main/java/net/silentchaos512/gear/api/item/GearType.java
@@ -223,6 +223,10 @@ public final class GearType {
                 .orElse(Collections.emptySet());
     }
 
+    public GearTypeMatcher getMatcher(boolean matchParents) {
+        return new GearTypeMatcher(matchParents, this);
+    }
+
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;