+++ src/main/java/net/silentchaos512/gear/data/part/PartsProvider.java
@@ -66,11 +66,18 @@ public class PartsProvider implements IDataProvider {
         });
 
         ret.add(upgradePart("misc/spoon", CraftingItems.SPOON_UPGRADE)
-                .upgradeGearTypes(new GearTypeMatcher(false, GearType.PICKAXE))
+                .upgradeGearTypes(GearType.PICKAXE.getMatcher(false))
                 .stat(ItemStats.DURABILITY, 0.2f, StatInstance.Operation.MUL1)
                 .stat(ItemStats.RARITY, 10, StatInstance.Operation.ADD)
                 .trait(Const.Traits.SPOON, 1)
                 .display(GearType.PICKAXE, new MaterialLayer(SilentGear.getId("spoon"), Color.VALUE_WHITE))
+        );
+        ret.add(upgradePart("misc/road_maker", CraftingItems.ROAD_MAKER_UPGRADE)
+                .upgradeGearTypes(GearType.EXCAVATOR.getMatcher(false))
+                .stat(ItemStats.DURABILITY, 0.1f, StatInstance.Operation.MUL1)
+                .stat(ItemStats.RARITY, 10, StatInstance.Operation.ADD)
+                .trait(Const.Traits.ROAD_MAKER, 1)
+                .display(GearType.EXCAVATOR, new MaterialLayer(SilentGear.getId("road_maker"), Color.VALUE_WHITE))
         );
         ret.add(upgradePart("misc/red_card", CraftingItems.RED_CARD_UPGRADE)
                 .upgradeGearTypes(GearTypeMatcher.ALL)