+++ src/main/java/net/silentchaos512/gear/gear/trait/TraitSerializers.java
@@ -42,6 +42,7 @@ public final class TraitSerializers {
         register(StatModifierTrait.SERIALIZER);
         register(AttributeTrait.SERIALIZER);
         register(BlockPlacerTrait.SERIALIZER);
+        register(BlockFillerTrait.SERIALIZER);
         register(SynergyTrait.SERIALIZER);
         register(TargetEffectTrait.SERIALIZER);
     }